feat(project): abstract download button
